refactor(arch): FSD cleanup — 의도적 예외 해소 8건 + cleanup 17건 + CLAUDE.md 문서화
의도적 예외 해소:
- useCurrentUser → entities/session/hooks/ 이동 (shared→entities 역방향 해소)
- cross-widget deep import 7건 → barrel import로 전환

Cleanup:
- Dead 파일 6건 삭제 (usePointerTooltip, useTrendlineOverlay, useKeyLevelsOverlay, usePatternOverlay, usageLogRepository, useSelectedProvider)
- Dead export 3건 제거
- 'use client' 누락 1건 수정 (NewsAiSummaryError.tsx)
- 빈 barrel 3건 정리
- cn.ts return type 추가

문서화:
- entities/CLAUDE.md, features/CLAUDE.md, widgets/CLAUDE.md 신규 작성 (의도적 예외 기록)
- shared/CLAUDE.md 갱신 (shared→entities 예외 기록)
