+++ src/entities/market-summary/index.ts
@@ -1,1 +1,2 @@
-// actions are imported from @/entities/market-summary/actions
+// actions are imported directly from @/entities/market-summary/actions
+export {};
