+++ src/entities/sector-signal/index.ts
@@ -1,1 +1,2 @@
-// actions are imported from @/entities/sector-signal/actions
+// actions are imported directly from @/entities/sector-signal/actions
+export {};
