+++ src/entities/session/hooks/useCurrentUser.ts
