+++ src/widgets/symbol-page/ChartContent.tsx
@@ -5,8 +5,8 @@ import React, { Suspense, useEffect, useEffectEvent, useMemo } from 'react';
 import dynamic from 'next/dynamic';
 import { type AnalysisResponse, type Timeframe } from '@y0ngha/siglens-core';
 import { cn } from '@/shared/lib/cn';
-import { ChartSkeleton } from '@/widgets/chart/ChartSkeleton';
-import { AnalysisPanel } from '@/widgets/analysis/AnalysisPanel';
+import { ChartSkeleton } from '@/widgets/chart';
+import { AnalysisPanel } from '@/widgets/analysis';
 import { BotBlockedNotice } from '@/shared/ui/BotBlockedNotice';
 import { useBars } from './hooks/useBars';
 import { useAnalysis } from './hooks/useAnalysis';
